Crypto-bot/requirements.txt
aiohttp==3.8.5
python-telegram-bot==20.6
telethon==1.28.5
python-dotenv==1.0.0
solana==0.30.2
requests==2.31.0 